pyproject.toml
[project]
name = "jax_generative_models"
version = "0.1.0"
description = "JAX implementation of generative models"
authors = [
  {name = "Mizuho Aoki", email = "mizuhoaoki1998@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.12, <4.0"

dependencies = [
    "equinox>=0.13.2",
    "jaxtyping>=0.3.3",
    "matplotlib>=3.10.7",
    "optax>=0.2.6",
    "optuna>=4.6.0",
    "qpax>=0.0.9",
    "rerun-sdk==0.23",
    "rich>=14.2.0",
    "scikit-learn>=1.7.2",
    "tyro>=0.9.35",
]

[project.optional-dependencies]
cpu = [
    "jax>=0.8.0",
]
cuda12 = [
    "jax[cuda12]>=0.8.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.5.4",
    "mypy>=1.10.0",
    "pre-commit>=4.3.0",
]

[tool.uv]
conflicts = [
  [
    { extra = "cpu" },
    { extra = "cuda12" },
  ],
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.mypy]
files = ["."]
exclude = ["outputs", "assets"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "optax"
ignore_missing_imports = true

[tool.ruff]
src = ["src", "scripts"]
line-length = 100

[tool.ruff.lint]
# Enable pycodestyle's E and W rules, and the isort rules
select = ["E", "W", "I", "F"]

[tool.ruff.format]
quote-style = "double"
